+++ src/structs/parameter_estimation.jl
@@ -62,6 +62,27 @@ function IpoptOptions(;
 end
 
 """
+    OptimisersOptions(; iterations::Integer = 100, max_time = Inf)
+
+Create options for `PEtab.calibrate` when using an `Optimisers.jl` optimizer rule.
+
+# Keyword Arguments
+- `iterations::Integer = 100`: Number of update steps to perform. Must be a positive
+   integer.
+- `max_time::AbstractFloat = Inf`: Maximum time in seconds that the optimization is
+   allowed to run.
+"""
+struct OptimisersOptions
+    iterations::Int64
+    max_time::Float64
+end
+function OptimisersOptions(; iterations::Integer = 100, max_time::Real = Inf)
+    @assert iterations > 0 "iterations must be a positive integer"
+    @assert max_time > 0 "max_time must be a positive number"
+    return OptimisersOptions(iterations, max_time)
+end
+
+"""
     PEtabOptimisationResult
 
 Parameter estimation statistics from single-start optimization with `calibrate`.